pyproject.toml
[tool.poetry]
name = "atmosphere"
version = "0.13.0"
description = ""
authors = ["Mohammed Naser <mnaser@vexxhost.com>"]
readme = "README.md"

[tool.poetry.scripts]
atmosphere = "atmosphere.cmd.cli:main"
atmosphere-operator = "atmosphere.cmd.operator:main"

[tool.poetry.dependencies]
python = "^3.10"
click = "^8.1.3"
pydantic = { version = "^1.10.4", optional = true }
schematics = { version = "^2.1.1", optional = true }
pykube-ng = { version = "^22.7.0", optional = true }
structlog = { version = "^22.1.0", optional = true }
mergedeep = { version = "^1.3.4", optional = true }
taskflow = { version = "^5.0.0", optional = true }
eventlet = { version = "^0.33.1", optional = true }
tomli = { version = "^2.0.1", optional = true }
jsonnet = { version = "^0.18.0", optional = true }
kopf = { version = "^1.36.0", optional = true, extras = ["uvloop"] }
openstacksdk = { version = "<0.99.0", optional = true }
certbuilder = { version = "^0.14.2", optional = true }
validators = { version = "^0.20.0", optional = true }
"oslo.log" = "^5.0.2"
"oslo.config" = "^9.0.0"
"oslo.concurrency" = "^5.0.1"
docker-image-py = "^0.1.12"

[tool.poetry.extras]
operator = [
  "schematics",
  "pydantic",
  "pykube-ng",
  "structlog",
  "rich",
  "better-exceptions",
  "mergedeep",
  "taskflow",
  "eventlet",
  "tomli",
  "jsonnet",
  "kopf",
  "openstacksdk",
  "certbuilder",
  "validators",
]

[tool.poetry.group.dev.dependencies]
ansible-core = "^2.13.4"
flake8 = "^5.0.4"
flake8-isort = "^4.2.0"
hypothesis = "6.61.0"
Jinja2 = "^3.1.2"
jinja2-ansible-filters = "^1.3.2"
jmespath = "^1.0.1"
molecule = "^4.0.1"
pytest = "^7.1.3"
pytest-cov = "^3.0.0"
pytest-kind = "^22.9.0"
pytest-mock = "^3.8.2"
python-on-whales = "^0.52.0"
tomli-w = "^1.0.0"
responses = "^0.22.0"

[tool.poetry.group.docs.dependencies]
mkdocs-material = "^8.5.7"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
profile = "black"

[tool.pytest.ini_options]
addopts = ["--cov=atmosphere", "--cov-report=term-missing"]
filterwarnings = [
  "ignore::DeprecationWarning",
  "ignore::schematics.deprecated.SchematicsDeprecationWarning",
  "ignore:The asyncore module is deprecated",
]

[tool.vulture]
paths = ["atmosphere"]
